requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
qdrant-client
llama-cpp-python
streamlit
python-dotenv
langchain
